Add the concepts of matrix_rule, simple_rule and RuleConfigs
This is the first step to be able to transform the metadata of GEO files.

This commit add a new kind of rule (simple_rule) and change the name of the transformation_rule to matrix_rule. The idea of this change is that the matrix rule will have the power to search a content by the column name (the same behavior of the former transformation_rule) and the simple_rule will deal only with the matrix cell that it is being processed. In the simple_rule the behavior of the software is different: the software will process no line-by-line but it will process each cell (line x column) and deal only with the current cell's content.

This is not a final commit about this new approach. This is the first step and did no change in the behavior yet.